requirements.txt
opencv-python
imagehash
Pillow